+++ Assets/NanoDetSample.cs
@@ -23,27 +23,24 @@ namespace NatML.Examples {
         private NanoDetPredictor predictor;
 
         private async void Start () {
-            // Fetch the model data from NatML Hub
-            modelData = await MLModelData.FromHub("@natsuite/nanodet");
-            // Create the model
-            model = new MLEdgeModel(modelData);
             // Create the NanoDet predictor
-            predictor = new NanoDetPredictor(model, modelData.labels);
+            predictor = await NanoDetPredictor.Create();
             // Listen for camera frames
-            cameraManager.OnFrame.AddListener(OnCameraFrame);
+            cameraManager.OnCameraFrame.AddListener(OnCameraFrame);
         }
 
         private void OnCameraFrame (CameraFrame frame) {
-            // Create input feature
-            var feature = frame.feature;
-            (feature.mean, feature.std) = modelData.normalization;
-            feature.aspectMode = modelData.aspectMode;
             // Detect
-            var detections = predictor.Predict(feature);
+            var detections = predictor.Predict(frame);
             // Visualize
             visualizer.Render(detections);
         }
 
-        void OnDisable () => model?.Dispose(); // Dispose the model
+        void OnDisable () {
+            // Stop listening for camera frames
+            cameraManager.OnCameraFrame.RemoveListener(OnCameraFrame);
+            // Dispose the model
+            model?.Dispose();
+        }
     }
 }