+++ Assets/NanoDetVisualizer.cs
@@ -8,6 +8,7 @@ namespace NatML.Examples.Visualizers {
     using System.Collections.Generic;
     using UnityEngine;
     using UnityEngine.UI;
+    using NatML.Vision;
     using VideoKit.UI;
 
     /// <summary>
@@ -22,28 +23,17 @@ namespace NatML.Examples.Visualizers {
 
         #region --Client API--
         /// <summary>
-        /// Detection source image.
-        /// </summary>
-        public Texture2D image {
-            get => rawImage.texture as Texture2D;
-            set {
-                rawImage.texture = value;
-                aspectFitter.aspectRatio = (float)value.width / value.height;
-            }
-        }
-
-        /// <summary>
         /// Render a set of object detections.
         /// </summary>
         /// <param name="image">Image which detections are made on.</param>
         /// <param name="detections">Detections to render.</param>
-        public void Render (params (Rect rect, string label, float score)[] detections) {
+        public void Render (params NanoDetPredictor.Detection[] detections) {
             // Delete current
             foreach (var rect in currentRects)
                 GameObject.Destroy(rect.gameObject);
             currentRects.Clear();
             // Render rects
-            var imageRect = new Rect(0, 0, image.width, image.height);
+            var imageRect = new Rect(0, 0, rawImage.texture.width, rawImage.texture.height);
             foreach (var detection in detections) {
                 var rect = Instantiate(detectionPrefab, transform);
                 rect.gameObject.SetActive(true);
@@ -56,13 +46,9 @@ namespace NatML.Examples.Visualizers {
 
         #region --Operations--
         private RawImage rawImage;
-        private AspectRatioFitter aspectFitter;
         private readonly List<NanoDetDetection> currentRects = new List<NanoDetDetection>();
 
-        void Awake () {
-            rawImage = GetComponent<RawImage>();
-            aspectFitter = GetComponent<AspectRatioFitter>();
-        }
+        void Awake () => rawImage = GetComponent<RawImage>();
         #endregion
     }
 }