+++ Packages/ai.natml.vision.nanodet/Changelog.md
@@ -1,2 +1,8 @@
+## 1.0.1
++ Added `NanoDetPredictor.Create` static method for creating the predictor.
++ Added `NanoDetPredictor.Tag` constant string for model embedding.
++ Removed `NanoDetPredictor` public constructor. Use the `Create` static method instead.
++ Upgraded to NatML 1.1.4.
+
 ## 1.0.0
 + First release.